Revert "sprint6"
This reverts commit 548406fe87ef598d7fa145eb4aa56563a3d8103c.
