+++ package.json
@@ -8,9 +8,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.24.0",
     "react-scripts": "5.0.1",
-    "styled-components": "^6.1.11",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
