+++ src/components/Button.js
@@ -1,15 +1,12 @@
 import './Button.css';
 
-
-
-
 function Button({ children, onClick, select = '' }) {
-  const classNames = `Button ${select}`;
-  return (
-    <button className={classNames} onClick={onClick}>
-      {children}
-    </button>
-  );
+    const classNames = `Button ${select}`;
+    return (
+        <button className={classNames} onClick={onClick}>
+            {children}
+        </button>
+    );
 }
 
 export default Button;
