--- src/components/FileInput.css
@@ -1,48 +1,0 @@
-.fileInputWrapper {
-  display: flex;
-  gap: 20px;
-}
-
-.fileInput {
-  display: none;
-}
-
-.fileInputButton {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  background-color: #f3f4f6;
-  width: 282px;
-  height: 282px;
-  border-radius: 12px;
-}
-
-.fileInputCustom {
-  color: #9ca3af;
-  font-size: 16px;
-  font-weight: 400px;
-}
-
-.imgPreviewWrapper {
-  position: relative;
-}
-
-.previewImg {
-  width: 282px;
-  height: 282px;
-  border-radius: 12px;
-}
-
-.imgRemoveButton {
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  background-color: #3692ff;
-  color: #ffffff;
-  border: none;
-  border-radius: 50%;
-  width: 30px;
-  height: 30px;
-  font-size: 20px;
-}
