+++ src/index.js
@@ -1,10 +1,10 @@
 import ReactDOM from 'react-dom/client';
-import Main from './Main';
+import App from "./components/App";
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
 
-  <Main></Main>
+  <App></App>
 
 
 );
