--- src/pages/Items.css
@@ -1,25 +1,0 @@
-.saleProductTitle {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.titleRight {
-  display: flex;
-  gap: 12px;
-}
-
-.productSearch {
-  width: 325px;
-  height: 42px;
-  border-radius: 12px;
-  background-color: #f3f4f6;
-  border: none;
-  color: #9ca3af;
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 24px;
-  padding: 9px 20px 9px 16px;
-  outline: none;
-}
