--- src/pages/NotFound.js
@@ -1,15 +1,0 @@
-
-
-function NotFound() {
-
-    return (
-        <>
-            <h1>??????</h1>
-            <h1>??????</h1>
-            <h1>??????페이지를 찾을수 없습니다</h1>
-        </>
-    )
-
-}
-
-export default NotFound;
